Stop the engine re-reading, re-hashing, and re-copying what it already has
Six wastes the second audit measured. The model child's output was
drained a byte at a time from an unbuffered pipe, so every byte of a
job's stdout cost a read syscall - 216,241 of them for a 216KB stream,
now 879 - and every tqdm redraw was forwarded to the log even when it
carried a percentage already reported, which is most of them. Buffering
the pipe and logging only a reading that actually changed cuts both the
syscalls and the volume the plugin then has to read and store.

MIDI export merged notes by walking the whole merged list per note,
which is quadratic and was already the hottest part of an export: 385s
at 64k notes, 0.16s now. The obvious pitch-to-newest-note index is not
equivalent - the old scan walks past a still-sounding note to reach an
older one whose end lands inside the gap window - so the index keeps a
per-pitch list, scanned newest first and pruned once entries fall out of
reach, and a test carries the case the naive version gets wrong. An
export that misses the cache also hashed the same stem twice; it hashes
once.

The Beat This! checkpoint was re-hashed in full on every resolution
even though the download had already verified it: a sidecar records the
size, mtime and digest that were checked, and only a file that no
longer matches is hashed again - 219-415ms per run down to 0.1ms, with
tests that a modified checkpoint is still caught and rejected. FLAC
input was decoded to float64 before being written back as 24-bit PCM,
doubling the buffer for no reason; reading at the width it writes leaves
the output bit-identical. Demucs staged its six stems under the system
temp root and copied them across filesystems to the job; the scratch
directory now opens beside the destination so they move instead - 997ms
of copying becomes 8.5ms - falling back to a copy when the rename
cannot cross.

+++ src/stemlab/pretrained.py
@@ -74,7 +74,13 @@ def _convert_with_soundfile(input_path: Path, staged: Path) -> bool:
     try:
         import soundfile as sf
 
-        data, sample_rate = sf.read(str(input_path), always_2d=True)
+        # Read at the width being written. Without a dtype soundfile decodes
+        # to float64 - eight bytes a sample, twice the peak this holds - for
+        # data that goes straight back out as PCM_24, which is what libsndfile
+        # produces from int32 by dropping the low eight bits. The whole track
+        # is in memory at once here, immediately before the separator loads
+        # its model, and the hybrid engine pays for it twice.
+        data, sample_rate = sf.read(str(input_path), always_2d=True, dtype="int32")
         sf.write(str(staged), data, sample_rate, subtype="PCM_24")
     except Exception:
         return False
